+++ @apicize/lib-typescript/src/models/validation.ts
@@ -15,6 +15,8 @@ export interface ValidationErrors {
  */
 export interface ValidationWarnings {
     validationWarnings?: string[]
+        validationState?: ValidationState
+
 }
 
 /**