+++ @apicize/toolkit/src/controls/navigation/lists/scenario-list.tsx
@@ -15,7 +15,7 @@ export const ScenarioList = observer(({ sx }: { sx?: SxProps, }) => {
     return <Box sx={sx} className='editor'>
         <Stack direction='row' className='editor-panel-header' flexGrow={0}>
             <EditorTitle icon={<SvgIcon color='scenario'><ScenarioIcon /></SvgIcon>} name='Scenarios'>
-                <IconButton color='primary' size='medium' aria-label='Close' title='Close' sx={{ marginLeft: '1rem' }} onClick={() => workspace.returnToNormal()}><CloseIcon fontSize='inherit' /></IconButton>
+                <IconButton color='primary' size='medium' aria-label='Close' title='Close' sx={{ marginLeft: '1rem' }} onClick={() => workspace.returnToNormal()}><CloseIcon fontSize='inherit' color='error' /></IconButton>
             </EditorTitle>
         </Stack>
         <Box className='editor-list'>