+++ @apicize/toolkit/src/models/workspace/execution.ts
@@ -18,7 +18,7 @@ export type InfoColorType = OverridableStringUnion<
 /**
  * Payload that application back-end notifies front-end of execution results
  */
-export type ExecutionEvent = ExecutionStartEvent | ExecutionCompleteEvent | ExecutionCancelEvent
+export type ExecutionEvent = ExecutionStartEvent | ExecutionCompleteEvent | ExecutionCancelEvent | ExecutionClearEvent
 
 export interface ExecutionStartEvent {
      eventType: 'start',
@@ -37,10 +37,17 @@ export interface ExecutionCancelEvent {
      executionState: ExecutionState
 }
 
+export interface ExecutionClearEvent {
+     eventType: 'clear',
+     executionState: ExecutionState
+     menu: ExecutionMenuItem[]
+     activeSummaries: { [execCtr: number]: ExecutionResultSummary }
+}
+
 export interface ExecutionMenuItem {
      name: string
      level: number
-     
+
      executingName?: string
      executionState: ExecutionState